pyproject.toml
[project]
name = "p2p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.116.1",
    "grpcio>=1.73.1",
    "grpcio-tools>=1.73.1",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
